PRO-4136 revising tests
