+++ test/component/summary/testSummary.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const TestWrapper = require('test/util/TestWrapper');
-const ThankYou = require('app/steps/ui/thankyou/index');
+const PaymentBreakdown = require('app/steps/ui/payment/breakdown');
 const nock = require('nock');
 const testHelpBlockContent = require('test/component/common/testHelpBlockContent.js');
 const minimalCaveatForm = require('test/data/unit/minimalcaveatform');
@@ -11,7 +11,7 @@ const security = require('app/components/security');
 
 describe('summary', () => {
     let testWrapper;
-    const expectedNextUrlForThankYou = ThankYou.getUrl();
+    const expectedNextUrlForThankYou = PaymentBreakdown.getUrl();
     let servicesMock, securityMock;
 
     beforeEach(() => {
@@ -65,57 +65,5 @@ describe('summary', () => {
 
     });
 
-    describe('sendToOrchestrationService()', () => {
-        it('should return to status 500 page when service authority fails', (done) => {
-            const sessionData = minimalCaveatForm;
-            const data = {};
-            servicesMock.expects('authorise').returns(Promise.resolve({
-                name: 'Error',
-                message: 'Unable to find service'
-            }));
-            testWrapper.agent.post('/prepare-session/form')
-                .send(sessionData)
-                .end((err) => {
-                    if (err) {
-                        throw err;
-                    }
-                    testWrapper.testStatus500Page(done, data);
-                });
-        });
-        it('should return to status 500 page when security idam calls fail', (done) => {
-            const sessionData = minimalCaveatForm;
-            const data = {};
-            servicesMock.expects('authorise').returns(Promise.resolve('authorised'));
-            securityMock.expects('getUserToken').returns(Promise.resolve({
-                name: 'Error',
-                message: 'Unable to find service'
-            }));
-            testWrapper.agent.post('/prepare-session/form')
-                .send(sessionData)
-                .end((err) => {
-                    if (err) {
-                        throw err;
-                    }
-                    testWrapper.testStatus500Page(done, data);
-                });
-        });
-        it('should return to status 500 page when orchestration call fails', (done) => {
-            const sessionData = minimalCaveatForm;
-            const data = {};
-            servicesMock.expects('authorise').returns(Promise.resolve('authorised'));
-            securityMock.expects('getUserToken').returns(Promise.resolve('token'));
-            servicesMock.expects('sendToOrchestrationService').returns(Promise.resolve({
-                name: 'Error',
-                message: 'Unable to find service'
-            }));
-            testWrapper.agent.post('/prepare-session/form')
-                .send(sessionData)
-                .end((err) => {
-                    if (err) {
-                        throw err;
-                    }
-                    testWrapper.testStatus500Page(done, data);
-                });
-        });
-    });
+
 });
